orchestration/requirements.txt
apache-beam[gcp]==2.46.0
apache-beam[interactive]==2.46.0
google-cloud-bigquery>=2.34.4
google-cloud-storage>=2.7.0
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0